add otcustomizers at module, update README customization docs
