+++ README.md
@@ -154,6 +154,23 @@ def run_custom_protocol(transfer_volume: float=1.0, number_of_rows: int=1):
 
 ### Experimental Customization Widgets
 
+In this repo, the `otcustomizers` python module contains some utilities that you can use to
+generate "customization widgets" like dropdown menus and file upload fields on your protocol's
+page on the Protocol Library.
+
+Unfortunately, the relative import for them is a little ugly.
+You will need to do this at the top of your protocol:
+
+```python
+from os import path, sys
+sys.path.append(path.abspath(path.join(
+    __file__, path.pardir, path.pardir, path.pardir)))  # noqa: E402
+```
+
+Now you can do `from otcustomizers import StringSelection, FileInput`, etc, and use
+these imports in your protocol.
+
+
 #### StringSelection
 
 A `StringSelection` argument to your `run_custom_protocol` function will create a dropdown menu
@@ -162,25 +179,12 @@ on your protocol's page on the Protocol Library website.
 To use it, first copy and paste this block near the top of your protocol:
 
 ```python
-class StringSelection(object):
-    def __init__(self, *containers):
-        self.accepted_containers = containers
+# Relative import hack
+from os import path, sys
+sys.path.append(path.abspath(path.join(
+    __file__, path.pardir, path.pardir, path.pardir)))  # noqa: E402
 
-    def generate_options(self):
-        def humanize(txt):
-            return txt.replace('-', ' ').replace('_', ' ')
-
-        return [
-            {'value': option, 'text': humanize(option)}
-            for option in self.accepted_containers]
-
-    def get_json(self):
-        # Of the form:
-        # {type: 'StringSelection',
-        # options: [{value: '96-flat', text: '96 flat'}, ...]}
-        return {
-            'type': 'StringSelection',
-            'options': self.generate_options()}
+from otcustomizers import StringSelection
 ```
 
 Then use it in your `run_custom_protocol` function:
@@ -206,8 +210,7 @@ And another selector menu is made for "Tuberack Type" with options: 'tube-rack-.
 #### FileInput
 
 A `FileInput` argument to your `run_custom_protocol` function creates a file upload button
-on your protocol's page on the Protocol Library website. To use it, copy and paste the `FileInput` class
-code (below).
+on your protocol's page on the Protocol Library website.
 
 The uploaded file will be passed into the `run_custom_protocol` function as plain text.
 
@@ -215,11 +218,12 @@ If you want to parse a `.csv` (comma separated values) text file, you can copy a
 the `well_csv_to_list` helper below.
 
 ```python
-class FileInput(object):
-    def get_json(self):
-        return {
-            'type': 'FileInput'
-        }
+# Relative import hack
+from os import path, sys
+sys.path.append(path.abspath(path.join(
+    __file__, path.pardir, path.pardir, path.pardir)))  # noqa: E402
+
+from otcustomizers import FileInput
 
 
 def well_csv_to_list(csv_string):
